Got rid of page template
search works again, only on home page. got rid of sign in button and menu button since neither of them did anything
