--- client/src/components/PageTemplate/PageTemplate.css
@@ -1,77 +1,0 @@
-.App {
-    text-align: center;
-}
-
-.header-container {
-    display: flex;
-    position: fixed; 
-    flex-direction: row;
-    height: 45px;
-    width:100vw;
-    align-items: center;
-    margin-left: 250px;
-    background-color: #6486A9;
-    z-index: 1;
-}
-
-.search-header {
-    height: 40px;
-    width: 545px;
-    display:inline-flex;
-    color:white;
-    text-align: left;
-    padding-top:1px;
-    padding-left:15px;
-}
-
-.search-header button {
-    display: inline-flex;
-    margin-left:10px;
-}
-
-.search-header button:hover {
-    background-color: #538cc6;
-}
-
-.menu-header {
-    color: black;
-    width: 50px;
-}
-
-.menu-header :hover {
-    background-color: #538cc6;
-}
-
-.right-header {
-    height: 40px;
-    width: 100%;
-    align-items: right;
-    padding-left:15%;
-}
-
-.right-header :hover {
-    background-color: #538cc6;
-    color:white;
-}
-
-.right-header img {
-    margin-left:8px;
-}
-
-.btn {
-    color:white;
-}
-
-.main-content {
-    overflow: auto;
-    margin-left: 250px;
-    text-align: left;
-    padding-top:60px;
-    padding-left:10px;
-    height: 100vh;
-}
-
-.App-link {
-    color: #61dafb;
-}
-
