+++ client/src/routes/HomePage/HomePage.js
@@ -1,6 +1,8 @@
 import { useState } from "react";
 import "./HomePage.css";
-import PageTemplate from "../../components/PageTemplate/PageTemplate";
+import list from "../../assets/list.png";
+import search from "../../assets/search.png";
+import signIn from "../../assets/signIn.png";
 import AssetTable from "../../components/AssetTable/AssetTable";
 import CatDropdown from "../../components/CatDropdown/CatDropdown";
 
@@ -9,20 +11,52 @@ export default function HomePage() {
         catName: undefined,
         category_id: -1
     })
-    const [inputValue, getInputValue] = useState({
-        inputVal: null
-    });
+
+    const [inputVal, setInputVal] = useState(null);
+
+    //Handling user input when user hits 'Enter'
+    function handleKeyPress(e) {
+        if (e.key === "Enter") {
+          console.log("Key press is entered");
+          getInputValue();
+        }
+      }
+    
+    function getInputValue() {
+        // Selecting the input element and get its value
+        const newInputVal = document.getElementById("search").value;
+        console.log("Input Value: " + newInputVal);
+        setInputVal(newInputVal);
+    };
 
     return (
         <div className="App">
-            <PageTemplate></PageTemplate>
+            <div className="header-container container-fluid">
+                <div className="search-header">
+                    <input
+                        type="text"
+                        onKeyDown={handleKeyPress}
+                        class="form-control"
+                        id="search"
+                        placeholder="Search"
+                        name="search"
+                    />
+                    <button type="submit" onClick={getInputValue} class="btn btn-default">
+                        <img src={search} alt="search" width="22" height="22" />
+                    </button>
+                    
+                </div>
+            </div>
+
             <div className="container-fluid main-content">
                 <div className="category">
                     <CatDropdown state={currentCategory} update={updateCategory} ></CatDropdown>
                 </div>
-                <AssetTable cat={currentCategory?.category_id} input={inputValue.inputVal}></AssetTable> 
+                <AssetTable cat={currentCategory?.category_id} input={inputVal}></AssetTable>
+
             </div> 
-    
+
         </div>
+
     );
 }
